+++ CHANGELOG.md
@@ -18,6 +18,7 @@ to [Semantic Versioning]. Full commit history is available in the
     AutoZI, CellAssign and GimVI. {pr}`3121`
 - Add {class}`scvi.external.RESOLVI` for bias correction in single-cell resolved spatial
     transcriptomics {pr}`3144`.
+- Add support for using Lamin custom dataloaders with {class}`scvi.model.SCVI`, {pr}`2932`.
 
 #### Fixed